__init__.py: FIX @app.route('/quizzes',methods=['POST']) ENDPOINT
Return Message when when there are no available questions.
Addresses:
quizzes. When we play a quiz questions are getting repeated for the same category, this should not happen as if we take the example of the Sports category, it contains two questions. After validating the two questions final score should be displayed but in your case question are getting repeated in the quiz. You should put check there when all questions are displayed from the category then the final score should be displayed.